+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add Google Site Verification to the bot list.
 - Handle invalid quality values that look like numbers.
 - Remove Browser::Bot.detect_empty_ua! in favor of Browser::Bot.allow_empty_ua!
+- Detect unknown platform/device as bots.
 
 ## v2.5.2